Update HSPipe.py
Had weird issues with trailing spaces after updating to dicts instead of array nightmares. Moving the trailing space check outside of the menu building loop seems to have fixed it. I don't particularly like having a whole loop specifically for that, but I don't want to mess with the dict size or popping entries in the middle of an interation. Just seems like a bad idea I guess.